requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
celery==5.3.4
python-dotenv==1.0.0
